Y4M loader: don't fail on newline
regression introduce in c08ae5f. if the color space was the last thing
in the header than it would be followed by a newline rather than space.

peek to ensure next char is either ' ' or newline to avoid falsly
matching problematic colorspaces. this will also work for other cases
like "mono-whatever" instead of working just for "420", "422" and "444".

also change the return code to unsupported instead of corrupted in these
cases.