Y4M loader: fix support for odd dimensions
Y4M adds an extra line for chromas when a file has an odd dimension. For
example, a 9x9 4:2:0 file has 81 pixels for luma, and 25 pixels for each
of the chromas (5 values for each of the 5 subsampled rows). Older code
was only reading 16 pixels (4x4), which caused the Cb chroma to be shifted,
and not enough pixels to be read.

Added a 9x9 y4m example, where the last column has a different chroma.